+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - OpenAI: eliminate use of `strict` tool calling (sporadically supported across models and we already interally validate).
 - Don't include package scope for task name part of log files.
 - Use user_data_dir rather than user_runtime_dir for view notifications.
+- Implement `read_eval_log_sample()` for JSON log files.
 
 ## v0.3.42 (23 October 2024)